@@ -2,14 +2,14 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Authors>Jonko</Authors>
     <Description>Wiinject is a tool for injecting ASM hacks into Wii games using Riivolution memory patches.</Description>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
     <RepositoryUrl>https://github.com/jonko0493/Wiinject</RepositoryUrl>
     <PackageProjectUrl>https://github.com/jonko0493/Wiinject</PackageProjectUrl>
     <RepositoryType>GitHub</RepositoryType>
-    <Version>1.2.1</Version>
+    <Version>1.2.2</Version>
   </PropertyGroup>
 
   <ItemGroup>
