@@ -4,8 +4,8 @@ namespace Wiinject
 {
     public class WiinjectResult
     {
-        public List<string> EmittedCFiles { get; set; } = new();
-        public Dictionary<string, byte[]> OutputBinaryPatches { get; set; } = new();
+        public List<string> EmittedCFiles { get; set; } = [];
+        public Dictionary<string, byte[]> OutputBinaryPatches { get; set; } = [];
         public Riivolution OutputRiivolution { get; set; }
     }
 }
