@@ -11,16 +11,18 @@ namespace Wiinject.Tests
         [TestCase("Forward", 0x80013040, 0x80012030, "48 00 0F F5")]
         public void ResolveBranchesTest(string direction, uint returnTestEntryPoint, uint callTestEntryPoint, string expectedAsm)
         {
-            List<CFunction> functions = new();
-            functions.Add(new("return_test", "00008000", TestHelpers.ReturnTestC) { EntryPoint = returnTestEntryPoint });
-            functions.Add(new("call_test", "00008000", TestHelpers.CallTestC) { EntryPoint = callTestEntryPoint });
+            List<CFunction> functions =
+            [
+                new("return_test", "00008000", TestHelpers.ReturnTestC) { EntryPoint = returnTestEntryPoint },
+                new("call_test", "00008000", TestHelpers.CallTestC) { EntryPoint = callTestEntryPoint },
+            ];
 
             functions[0].ResolveFunctionRefs(functions);
             functions[1].ResolveFunctionRefs(functions);
 
-            List<CFunction> resolvedFunctions = new();
+            List<CFunction> resolvedFunctions = [];
             resolvedFunctions = functions[1].FunctionsToResolve(resolvedFunctions);
-            Assert.AreEqual(functions, resolvedFunctions);
+            Assert.That(functions, Is.EquivalentTo(resolvedFunctions));
 
             functions[0].ResolveBranches();
             functions[1].ResolveBranches();
@@ -29,25 +31,24 @@ namespace Wiinject.Tests
             functions[1].SetDataFromInstructions();
 
             Console.WriteLine(functions[1].Data.ToHexString());
-            Assert.IsTrue(functions[1].Data.ToHexString().Contains(expectedAsm));
+            Assert.That(functions[1].Data.ToHexString(), Contains.Substring(expectedAsm));
         }
 
         [Test]
         public void ResolveRecursiveBranchTest()
         {
-            List<CFunction> functions = new();
-            functions.Add(new("recursion_test", "00008000", TestHelpers.RecursionTestC) { EntryPoint = 0x8001200 });
+            List<CFunction> functions = [new("recursion_test", "00008000", TestHelpers.RecursionTestC) { EntryPoint = 0x8001200 }];
             functions[0].ResolveFunctionRefs(functions);
 
-            List<CFunction> resolvedFunctions = new();
+            List<CFunction> resolvedFunctions = [];
             resolvedFunctions = functions[0].FunctionsToResolve(resolvedFunctions);
-            Assert.AreEqual(functions, resolvedFunctions);
+            Assert.That(functions, Is.EquivalentTo(resolvedFunctions));
 
             functions[0].ResolveBranches();
             functions[0].SetDataFromInstructions();
 
             Console.WriteLine(functions[0].Data.ToHexString());
-            Assert.IsTrue(functions[0].Data.ToHexString().Contains("4B FF FF D1"));
+            Assert.That(functions[0].Data.ToHexString(), Contains.Substring("4B FF FF D1"));
         }
     }
 }
